--- a/Proofs/AES-GCM/GCMGmultV8Sym.lean
+++ b/Proofs/AES-GCM/GCMGmultV8Sym.lean
@@ -119,11 +119,11 @@ theorem gcm_gmult_v8_program_run_27 (s0 sf : ArmState)
     simp_mem; rfl
     -/
     rw [Memory.read_bytes_write_bytes_eq_read_bytes_of_mem_separate']
-    simp_mem
+    mem_omega!
   · simp only [List.mem_cons, List.mem_singleton, not_or, and_imp]
     sym_aggregate
   · intro n addr h_separate
-    simp_mem (config := { useOmegaToClose := false })
+    simp_mem
     -- Aggregate the memory (non)effects.
     simp only [*]
   · clear_named [h_s, stepi_]
@@ -135,7 +135,14 @@ theorem gcm_gmult_v8_program_run_27 (s0 sf : ArmState)
       rw [@Memory.read_bytes_eq_extractLsBytes_sub_of_mem_subset'
            32 (r (StateField.GPR 1#5) s0) HTable (r (StateField.GPR 1#5) s0) 16 _ h_HTable.symm]
       · simp only [Nat.reduceMul, BitVec.extractLsBytes, Nat.sub_self, Nat.zero_mul]
-      · simp_mem
+      · mem_omega!
+        -- have := h_mem_sep.get 0 1
+        -- simp at this
+        -- rw [mem_separate'.iff_omega] at this
+        -- mem_omega!
+        -- mem_omega!
+        -- simp_all
+
     have h_HTable_high :
       (Memory.read_bytes 16 (r (StateField.GPR 1#5) s0 + 16#64) s0.mem) = HTable.extractLsb' 128 128 := by
       -- (FIXME @bollu) use `simp_mem` instead of the rw below.